Add a helper native library, "unloadPreloadedConflictingNativeLibs", to the Android sample framework project. It contains a function for unloading a set of native libraries in an Android application.
This is useful in the following scenario: say that our application needs to use a version of e.g. the PCRE library built especially for it. Now say, that Android already contains a system version of this library (e.g. "/system/lib/libpcre.so"), and say that this system version is already loaded by the application. If we don't unload the system version of PCRE, then our version (which is included in the APK) will not be used.

So at the beginning of our Android activity we load this helper library (the one that contains this source file), and use the following function to unload any version of any native library included in our APK, which is not the version included in our APK. We can find the set of full paths to the loaded libraries by examining the file "/proc/self/maps".