fix uninitialized variable
in a loop in GridLayoutContainer, two loop variables x,y are
not initialized except in the for loop body. However, the inner
loop may not run if the outer loop is skipped, and then y is
returned uninitialized.

Also fixes some unused var warnings which occur in release builds
when asserts are stripped out.